src/otonom/inference.py
from __future__ import annotations

from dataclasses import dataclass
from io import BytesIO
import math
from pathlib import Path
from random import Random
from typing import Any

import numpy as np
from PIL import Image

from .models import Detection


@dataclass(slots=True)
class InferenceOutput:
    backend: str
    detections: list[Detection]


class YOLOModelRuntime:
    """Loads TensorRT first, then ONNX, then deterministic fallback."""

    def __init__(
        self,
        model_path: str | None,
        tensorrt_engine_path: str | None = None,
        backend_preference: str = "tensorrt",
        seed: int = 42,
    ) -> None:
        self.model_path = Path(model_path) if model_path else None
        self.tensorrt_engine_path = Path(tensorrt_engine_path) if tensorrt_engine_path else None
        self.backend_preference = backend_preference.lower()
        self._rng = Random(seed)
        self._backend_name = "deterministic-fallback"
        self._session: Any = None
        self._input_name: str | None = None
        self._load_backend()

    @property
    def backend_name(self) -> str:
        return self._backend_name

    def _load_backend(self) -> None:
        if self.backend_preference == "onnx":
            if self._try_load_onnx():
                return
            if self._try_load_tensorrt():
                return
            return

        if self._try_load_tensorrt():
            return
        if self._try_load_onnx():
            return

    def _try_load_tensorrt(self) -> bool:
        if not self.tensorrt_engine_path or not self.tensorrt_engine_path.exists():
            return False
        try:
            import tensorrt as _  # type: ignore  # noqa: F401
            import pycuda.driver as _  # type: ignore  # noqa: F401
            self._backend_name = "tensorrt"
            return True
        except Exception:
            return False

    def _try_load_onnx(self) -> bool:
        if not self.model_path or not self.model_path.exists():
            return False
        try:
            import onnxruntime as ort

            self._session = ort.InferenceSession(str(self.model_path), providers=["CPUExecutionProvider"])
            self._input_name = self._session.get_inputs()[0].name
            self._backend_name = "onnxruntime"
            return True
        except Exception:
            self._session = None
            self._input_name = None
            return False

    def detect_from_image_bytes(self, image_bytes: bytes) -> InferenceOutput:
        image = Image.open(BytesIO(image_bytes)).convert("RGB")
        if self._backend_name == "onnxruntime" and self._session and self._input_name:
            detections = self._infer_onnx(image)
            return InferenceOutput(backend=self._backend_name, detections=detections)

        detections = self._infer_fallback(image)
        return InferenceOutput(backend=self._backend_name, detections=detections)

    def _infer_onnx(self, image: Image.Image) -> list[Detection]:
        tensor = self._preprocess(image)
        outputs = self._session.run(None, {self._input_name: tensor})
        if not outputs:
            return self._infer_fallback(image)

        raw = np.asarray(outputs[0])
        return self._decode_yolo_output(raw)

    def _preprocess(self, image: Image.Image, size: int = 640) -> np.ndarray:
        resized = image.resize((size, size))
        arr = np.asarray(resized).astype(np.float32) / 255.0
        arr = np.transpose(arr, (2, 0, 1))
        return np.expand_dims(arr, axis=0)

    def _decode_yolo_output(self, output: np.ndarray, conf_threshold: float = 0.75) -> list[Detection]:
        # Supports common YOLO ONNX outputs: (1, N, 84) or (1, 84, N)
        if output.ndim != 3:
            return []

        if output.shape[1] < output.shape[2]:
            pred = output[0]
        else:
            pred = output[0].transpose(1, 0)

        detections: list[Detection] = []
        for idx, row in enumerate(pred[:400]):
            if row.shape[0] < 5:
                continue
            x, y, w, h = row[:4]
            class_scores = row[4:]
            if class_scores.size == 0:
                continue
            raw_score = float(np.max(class_scores))
            conf = 1.0 / (1.0 + math.exp(-raw_score))
            if conf < conf_threshold:
                continue

            cx = float(np.clip(x / 640.0, 0.0, 1.0))
            cy = float(np.clip(y / 640.0, 0.0, 1.0))
            _ = w, h
            if cx < 0.001 and cy < 0.001:
                continue
            detections.append(
                Detection(
                    id=f"img-d{idx}",
                    label="zararli_bitki",
                    confidence=round(conf, 3),
                    image_x=round(cx, 3),
                    image_y=round(cy, 3),
                )
            )

            if len(detections) >= 30:
                break

        if not detections:
            return self._infer_fallback(Image.new("RGB", (640, 640), "white"))
        return detections

    def _infer_fallback(self, image: Image.Image) -> list[Detection]:
        count = 2 + self._rng.randint(0, 2)
        detections: list[Detection] = []
        for idx in range(count):
            conf = round(0.58 + self._rng.random() * 0.4, 3)
            detections.append(
                Detection(
                    id=f"fallback-d{idx}",
                    label="zararli_bitki",
                    confidence=conf,
                    image_x=round(0.2 + (idx + 1) / (count + 1) * 0.6, 3),
                    image_y=round(0.25 + self._rng.random() * 0.5, 3),
                )
            )
        return detections
