pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otonom"
version = "0.1.0"
description = "Safe autonomous weed detection and intervention mission simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "python-multipart>=0.0.9",
  "numpy>=1.26.0",
  "Pillow>=10.4.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6.0",
  "httpx>=0.27.0",
  "onnxruntime>=1.20.0"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
